Moving assets logic back to 'deploy' hook.
This undoes the back-and-forth commits cf6040b9, 1059c92, and 443efee,
which I believe has caused the asset pipeline logic to fail:
https://bugzilla.redhat.com/show_bug.cgi?id=962432

The reasons for the above commits are not documented anywhere, and I
cannot remember, either. If there is an issue arising from this, we will
document it the next time.