Effectively revert 6d06a0b.
The intent of the commit 6d06a0b was to speed up the deployment, but due
to the platform constraints, it broke scaled applications, where the
secondary gears do not have assets in the correct place. (The assets are
not copied from the primary gear to the secondary ones because the
assets were stored in $OPENSHIFT_DATA_DIR by this commit, so symbolic
links, no matter how they are to be created, are ineffective.) This led
to issues such as https://bugzilla.redhat.com/show_bug.cgi?id=962432.

This commit also fixes
https://bugzilla.redhat.com/show_bug.cgi?id=968248, which was triggered
by my previous attempt to make it all work together.